sun50i-codec-analog: add binding for internal bias resistor
In order to properly bias headset microphones, there should be a
resistor between pins HBIAS and MICDET. This can be an external
resistor, but the codec also provides an internal 2.2K resistor which
can be enabled by a register.

This patch adds a device-tree property and the corresponding processing
to the sun50i-codec-analog driver to take advantage of this feature.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>