sun50i-analog-codec: add binding for digital codec
In order to be able to enable jack detection, this analog codec must be
aware of the corresponding digital codec. This is achieved by
introducing a new required device-tree property.

This property is then used to retrieve a pointer to the private data of
the codec, which we'll use later on for enabling jack detection.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>