[WIP] pinctrl PM for A64 EMAC/I2C/MMC/SPI
Measurements taken on Orange Pi Win.

Impact on idle power consumption is around -0.1% or +0.1%, depending
on the run, but consistent within each run.

Using the "disable" function appears to be better than "gpio_in", but I
do not have enough data for "gpio_in" to say by how much.

In any case, the difference is at most 1 mW when idle, and 0.3 mW when
asleep, so pinctrl PM would be a very low priority.

Signed-off-by: Samuel Holland <samuel@sholland.org>